requirements.txt
appnope==0.1.3
backcall==0.2.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
debugpy==1.6.7
decorator==5.1.1
entrypoints==0.4
filelock==3.12.2
Flask==2.2.5
fsspec==2023.1.0
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.7.0
ipykernel==6.16.2
ipython==7.34.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jupyter-client==7.4.9
jupyter-core==4.12.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.7
numpy==1.21.6
packaging==23.1
pandas==1.3.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pyarrow==12.0.1
Pygments==2.15.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.0
regex==2023.6.3
requests==2.31.0
safetensors==0.3.1
six==1.16.0
tokenizers==0.13.3
torch==1.13.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.2
typing-extensions==4.7.1
urllib3==2.0.4
wcwidth==0.2.6
Werkzeug==2.2.3
zipp==3.15.0
